print the loss

+++ data/Transform_custom.py
@@ -183,6 +183,6 @@ class GaussianNoiseTensor():
         image_array = image_clone.numpy()
         #noise_img = random_noise(image_array, var=self.variance)
         #noise_img = torch.from_numpy(noise_img).float()
-        noise_img = torch.from_numpy(image_array).float()
+        noise_img = torch.from_numpy(image_array)
         return noise_img
